fix-quiz-information-in-separated-component
